Re-factor to allow chainbreaker to be used as module
