+++ chainbreaker/command_line.py
@@ -1,0 +1,4 @@
+import chainbreaker
+
+def main():
+    chainbreaker.main()
