+++ setup.py
@@ -22,5 +22,8 @@ setup(
         'Programming Language :: Python :: 3.7',
         'Programming Language :: Python :: 3.8',
     ],
+    entry_points = {
+        'console_scripts': ['chainbreaker=chainbreaker.command_line:main'],
+    },
     install_requires=dependencies,
 )
